backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg[binary]==3.2.1
cryptography==43.0.1
gunicorn==22.0.0
